Update test_math expectations for numpy+floatX dtypes
Comparison constants from Python int literals are now int64 (not int8) under
the new default cast_policy. Update the hardcoded expected dtypes in the
TestExpLog log/log1p/log1mexp switch tests and the TestSqrSqrt switch tests.

+++ tests/tensor/rewriting/test_math.py
@@ -1982,7 +1982,7 @@ class TestExpLog:
             f.maker.fgraph.outputs,
             [
                 pt.switch(
-                    x >= np.array([[0]], dtype=np.int8),
+                    x >= np.array([[0]], dtype=np.int64),
                     pt.log1p(x),
                     np.array([[np.nan]], dtype=np.float32),
                 )
@@ -2007,7 +2007,7 @@ class TestExpLog:
             f.maker.fgraph.outputs,
             [
                 pt.switch(
-                    x >= np.array([[0]], dtype=np.int8),
+                    x >= np.array([[0]], dtype=np.int64),
                     pt.log1p(-x),
                     np.array([[np.nan]], dtype=np.float32),
                 )
@@ -2030,7 +2030,7 @@ class TestExpLog:
             f.maker.fgraph.outputs,
             [
                 pt.switch(
-                    x <= np.array([[0]], dtype=np.int8),
+                    x <= np.array([[0]], dtype=np.int64),
                     x,
                     np.array([[np.nan]], dtype=np.float32),
                 )
@@ -2089,7 +2089,7 @@ class TestSqrSqrt:
         out = rewrite_graph(out, include=["canonicalize", "specialize", "stabilize"])
 
         expected = switch(
-            ge(x, np.zeros((1, 1), dtype="int8")),
+            ge(x, np.zeros((1, 1), dtype="int64")),
             x,
             np.full((1, 1), np.nan, dtype=out.type.dtype),
         )
@@ -2111,7 +2111,7 @@ class TestSqrSqrt:
         out = rewrite_graph(out, include=["canonicalize", "specialize", "stabilize"])
 
         expected = switch(
-            ge(x, np.zeros((1,), dtype="int8")),
+            ge(x, np.zeros((1,), dtype="int64")),
             x,
             np.full((1,), np.nan, dtype=dtype),
         )
